+++ CHANGELOG.md
@@ -10,7 +10,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] (2.0.0)
+## [2.0.0] - 2026-09-18
 ### Changed
 - Copyright notices are now represented by `CopyrightNotice` class instead of raw strings.
 
@@ -35,4 +35,5 @@ This is the first published version of the package. It doesn't contain any featu
 [0.0.0]: https://github.com/ForNeVeR/dotnet-reuse/releases/tag/v0.0.0
 [1.0.0]: https://github.com/ForNeVeR/dotnet-reuse/compare/v0.0.0...v1.0.0
 [1.1.0]: https://github.com/ForNeVeR/dotnet-reuse/compare/v1.0.0...v1.1.0
-[Unreleased]: https://github.com/ForNeVeR/dotnet-reuse/compare/v1.1.0...HEAD
+[2.0.0]: https://github.com/ForNeVeR/dotnet-reuse/compare/v1.1.0...v2.0.0
+[Unreleased]: https://github.com/ForNeVeR/dotnet-reuse/compare/v2.0.0...HEAD